[SCRIPT] Check for qemu by command -v

+++ build64/debug_memtest.sh
@@ -1,33 +1,35 @@
 #! /bin/bash
 
-#########################
-##
-##	TODO: DESCRIPTION
-##
 
-Info() {
-	echo "Error: Invalid option"
-	echo "Type $0 -h for more information"
-}
+###############################################################################
+#
+# Description	: `debug_memtest.sh` is a script that allows memtest86plus 
+#		developers to set up a debugging environment with GDB in QEMU. It calls
+#		the `make debug` target of a modified Makefile to create an additional
+#		debug-symbol file called `memtest.debug`. The symbols of this file are
+#		loaded with an offset (in accordance with the loading location of the
+#		efi-image) into GDB to match the exact addresses of the symbols in memory.
+#
+#		For more detailed information, please read 'HOW_TO_DEBUG_WITH_GDB'
+#
+# Author       	: Regina König
+# Year			: 2022
+# Email        	: koenig_regina@arcor.de
+#
+##############################################################################
 
 Help() {
-	echo "TODO Add description..."
-	echo
-	echo "Syntax: $0 [-h|c|t <terminal><execution_command>]"
+	echo "Syntax: $0 [-h|-c|-t <terminal><execution_command>]"
 	echo "options:"
-	echo "h	print this help"
-	echo "c	delete all debugging related files"
-	echo "t define alternative terminal"
-	echo
-}
-
-Terminal_Help() {
-	echo "You can define your own terminal inclusive its execution command via:"
-	echo "./debug_script.sh -t \"<terminal> <execution_command> \""
-	echo "See following examples:"
-	echo "./debug_script.sh -t \"x-terminal-emulator -e \""
-	echo "./debug_script.sh -t \"gnome-terminal --  \""
-    echo ".debug_script.sh -t \"xterm -e \""
+	echo " -h	Print this help"
+	echo " -c	Delete all debugging related files"
+	echo " -t	Define an alternative terminal"
+	echo "	You can define your own terminal inclusive its execution command via:"
+	echo "		./debug_script.sh -t \"<terminal> <execution_command> \""
+	echo "	See following examples:"
+	echo "		./debug_script.sh -t \"x-terminal-emulator -e \""
+	echo "		./debug_script.sh -t \"gnome-terminal --  \""
+    echo "		./debug_script.sh -t \"xterm -e \""
 }
 
 Clear() {
@@ -53,34 +55,32 @@ while getopts ":hct:" option; do
 			Clear
 			exit;;
 		t) # define own terminal
-			echo "argument -t called with parameter $OPTARG" >&2
-            if [ $OPTARG = "help" ]; then
-                Terminal_Help
-                exit 0
-            fi
 			TERMINAL="$OPTARG"
 			if ! $TERMINAL ls; then
 				echo "Your entered command is not valid. Please check it again"
-                echo "Or type \"./debug_memtest.sh -t help\" for help"
+                echo "Or type \"./debug_memtest.sh -h\" for help"
 				exit 1
 			fi
 			exit;;
         \?) # invalid option
-            Info
+            echo "Error: Invalid option"
+			echo "Type $0 -h for more information"
             exit;;
 	esac
 
 done
 
 Check() {
-	# Check if QEMU and OVMF is installed
-	if ! dpkg -l qemu-system* > /dev/null 2>&1; then
+	# Check if QEMU and OVMF are installed
+	if ! command -v qemu-system-x86_64 > /dev/null 2>&1; then
 		echo "Qemu not installed"
 		exit 1
 	fi
 
-	if ! dpkg -l ovmf > /dev/null 2>&1; then
-		echo "Package ovmf not installed. Type 'sudo apt install ovmf'"
+	# Check for presence of OVMF.fd, OVMF_VARS.fd and OVMF_CODE.fd
+	if [ ! -f OVMF.fd ] && [ ! -f /usr/share/ovmf/OVMF.fd ]; then
+		echo "Package ovmf not installed. Type 'sudo apt install ovmf'."
+		echo "Or copy your own versions of OVMF.fd, OVMF_VARS.fd and OVMF_CODE.fd into this directory"
 		exit 1
 	fi
 
@@ -98,7 +98,7 @@ Check() {
 		else
             echo "No terminal recognized. Please install x-terminal-emulator or gnome-terminal or xterm."
             echo "Or define your own terminal alternatively."
-			Terminal_Help
+			echo "Type ./debug_memtest.sh -h for more information"
 			exit 1
 		fi
 	fi
@@ -131,13 +131,27 @@ Init() {
 	QEMU_FLAGS+=" -drive if=pflash,format=raw,readonly=on,file=OVMF_CODE.fd"
 	QEMU_FLAGS+=" -drive if=pflash,format=raw,file=OVMF_VARS.fd"
 
+	# Define offsets for loading of symbol-table
+	IMAGEBASE=0x200000
+	BASEOFCODE=0x1000
+	RELOCADDR=0x400000
+
+	printf -v OFFSET "0x%X" $(($IMAGEBASE + $BASEOFCODE))
+
+
 	GDB_FILE="gdbscript"
-    # TODO Check if gdbscript exists - if yes do not create it
+
+	# Check if gdbscript exists. If not, create one.
     if [ ! -f $GDB_FILE ]
     then
         echo "Creating gdbscript.."
 
         echo "set pagination off" > $GDB_FILE
+
+        if [ -z "$OFFSET" ] || [ -z "$RELOCADDR" ]; then
+        	echo "OFFSET or RELOCADDR is not set."
+        	exit 1
+        fi
 
         echo "add-symbol-file memtest.debug $OFFSET" >> $GDB_FILE
         echo "add-symbol-file memtest.debug $RELOCADDR" >> $GDB_FILE
@@ -150,26 +164,33 @@ Init() {
 
         echo "b run_at" >> $GDB_FILE
 
-        echo "shell sleep 0.5" >> $GDB_FILE # TODO remove if time is enough until qemu starts
+        echo "shell sleep 0.5" >> $GDB_FILE
         echo "target remote localhost:1234" >> $GDB_FILE
         echo "info b" >> $GDB_FILE
         echo "c" >> $GDB_FILE
     fi
 
-	# Define offsets for loading of symbol-table
-	IMAGEBASE=0x200000
-	BASEOFCODE=0x1000
-	RELOCADDR=0x400000
-
-	printf -v OFFSET "0x%X" $(($IMAGEBASE + $BASEOFCODE))
+	if [ ! -f ldscripts/memtest_shared.lds ]; then
+		echo "'memtest_shared.lds' does not exist."
+		exit 1
+	fi
 
 	sed '/DISCARD/d' ldscripts/memtest_shared.lds > memtest_shared_debug.lds
 
+	if [ ! -f memtest_shared_debug.lds ]; then
+		echo "Creation of 'memtest_shared_debug.lds' failed."
+		exit 1
+	fi
 }
 
 Prepare_Directory() {
-    # Create dir hda-contents
+    # Create dir hda-contents and a boot directory
     mkdir -p hda-contents/EFI/boot
+
+    if [ ! -d hda-contents ]; then
+    	echo "Creation of directoeyy hda-contents failed."
+    	exit 1
+    fi
 
     # Copy memtest.efi to hda-contents
     cp memtest.efi hda-contents/
